Fix introduced address issue
