+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/types/address.d.ts
@@ -1,0 +1,4 @@
+interface AddressesData {
+  residential: Components.Schemas.Address;
+  mailing: Components.Schemas.Address;
+}
