+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/utils/constant.ts
@@ -15,3 +15,8 @@ export enum AlertNotificationType {
   SUCCESS = "success",
   INFO = "primary",
 }
+
+export enum AddressType {
+  RESIDENTIAL = "residential",
+  MAILING = "mailing",
+}
